Global SOLO implemented
The SOLO button in the Transport section now toggles soloing for multiple tracks simultaneously.

If one or more tracks are soloed, the SOLO button in the Transport section lights up. Press it and all tracks are unsoloed, but the button keeps blinking to indicate it has stored the previously soloed tracks. Press it again and those tracks are soloed again. Useful for toggling between all tracks and a soloed selection of tracks.

SOLO to toggle soloing for soloed tracks.
SHIFT + SOLO to reset.

The SOLO button's previous function, swap Device/Clip View, is moved to AUX button (a more appropriate location, next to the AUDIO INST button, which shows/hides Device/Clip View).