Macro Mapper beta
New feature: Macro Mapper. Experimental, see below for things to test. The purpose of this feature is to offer 16 user-definable controls, each mappable to multiple parameters anywhere in Live and controlled by the rotary knobs, which can be enabled by pressing the USER button.

Included in the folder XTouch/Macro Mapper/ is a Rack device called "X-Touch Macro Mapper.adg". Add it to your Master Track as the first (leftmost) device in the chain.

The Rack has 16 macros, each controlling an instance of a Max4Live device called "X-Touch MultiMap.amxd", which is also included in the same folder. Each of those 16 Max4Live devices can in turn be mapped to 8 different parameters anywhere in Live. Click Show Devices, then unfold any of the 16 X-Touch MultiMap devices to map them to the parameters you want to control.

In this way, each of the 16 macro knobs of the Rack can directly control up to 8 parameters in your project. You can set them up any way you want, and rename each of the 16 macros as you see fit.

* On your X-Touch unit:

If you press and hold the USER button, the rotary encoders and their scribble strips will be temporarily assigned to those 16 macro knobs (or the first 8 if you have only one unit; in that case, press INST/EQ to bank right/left, just like in Plugin Mode). Release the USER button to return to normal operation. SHIFT + USER to switch to the macro knobs permanently (until USER is pressed again).

If the "X-Touch Macro Mapper.adg" rack device is not found as the first (leftmost) device in the Master chain, the button will simply blink.

Things to test: drag the "X-Touch Macro Mapper.adg" device into your master track. Click Show Devices. Are the 16 "X-Touch MultiMap.amxd" Max4Live devices correctly included? I'm not sure if the link will be preserved when sharing in this way. Please report if not.