Chains behave as groups in mixer
- New: chains now behave as groups.

If a track contains a rack device with chains it can show in the mixer, those chains can be unfolded an get their own channel strips on the X-Touch just like tracks in a group. This means you can mix the individual chains in a rack (e.g., the individual sounds in a drum machine) along with the rest of your song. Chains can be muted and soloed (and will be included in Global Solo). Chain colors behave just like track colors (show up on scribble strips and can be changed from the X-Touch). Chain sends (to Return Chains within the same rack device) can be controlled in default Sends mode (select chain, press SEND; will default to the parent track's global sends if the device contains no Return Chains).

Limitations:
- Chains audio will not show up on the X-Touch meters, since the API does not provide access to chains audio levels.
- Chains "Audio To" cannot be controlled from the X-Touch like it can with normal tracks, since the API does not provide access to chains audio routing.
- Single Send mode will skip chains, since they have no sends to the global Return Tracks.

- New: for both chains and normal groups, the SELECT button of the parent track will now blink if a chain/track inside it is selected, to indicate at a glance where the current group/chain set starts.