Colors in default Sends mode + Single Send mode handling with chains
- In default Sends mode (controlling all available sends for a single track), the scribble strips now get the colors of the corresponding Return Tracks (or Return Chains, if a chain is currently selected).
- In Single Send mode, sends for chains are now also shown. Note that these are sends to Return Chains within the device, not to global Return Tracks. They will show the color of the corresponding Return Chain.

Current restriction: the number of available global Return Tracks is used to set the limit for banking between sends using the EQ and INST buttons. If a device with chains has more Return Chains than there are global Return Tracks in the song, some of them will not be viewable in Single Send mode.