Working on Macro Mapper mode for F1-F8
SHIFT + F8 sets Macro Mapper mode for the F1-F8 buttons. Press an unlit button to save the current state of the Macro Mapper parameters. Press a lit button to recall the saved state associated with it. Press OPTION + lit F1-F8 button to delete saved state.

Access to variations through the API is limited: we can't retrieve their names or IDs, Live doesn't tell us which one is active, all we know is how many there are. For that reason, we can't bind a variation to a specific F1-F8 button: if you delete the variation in F3, all variations above it will shift one to the left.