Option: moving fader requires finger touch
Added: option in settings menu to require finger (capacitive) touch for fader control (prevents accidental change of track volumes).

When this option is enabled, moving the fader with a non-capacitive object will not change the track volume, and the fader will jump back to its original location after a short time.